requirements.txt
blinker==1.9.0
click==8.1.8
python-dotenv==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
pytz==2024.1

Flask>=3.0
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
Flask-Migrate>=4.0
flask_restful>=1.0.0

# Markdown + Meta parsing
Markdown>=3.5
PyYAML>=6.0
python-frontmatter>=1.0.0

# FlatPages
Flask-FlatPages>=0.7

# Optionnel mais utile pour rechargement automatic
watchdog>=4.0.0

# HTTP client
requests>=2.31.0
